% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_NDC.R
\docType{data}
\name{datestamp}
\alias{datestamp}
\title{Read in a query from a GCAM run and establish NDC constraints with the fed in emissions as
a baseline.}
\format{An object of class \code{character} of length 1.}
\usage{
datestamp
}
\arguments{
\item{NDC_QUERY}{The globally averaged data}
}
\value{
Set of XML constraint/linking files
argparser argparser add_argument parse_args parse_args
}
\description{
The input should be .csv results of the query within the 
package (/data/queries/batch_query_NDC.xml).
}
\details{
The output will be a set of XMLs with varying constraints and linking files:
}
\keyword{datasets}
